[DF] Allow building clustered ranges locally or remotely
This commit introduces a new logic to build partitions of a distributed RDataFrame. On the client machine, the number of files of the dataset is split into percentages. Each partition will be sent to a worker that will then convert the percentage into an actual range of cluster boundaries in the file.

The new logic can be activated by setting

```
ROOT.RDF.Experimental.Distributed.open_files_locally = False
```