[df] Support DefinePerSample for RNTuple
The information of each sample is populated at every new `GetEntryRanges` call.
For the single thread case, every call corresponds to a new file. For the MT
case, different ranges might be accessing the same file. Extra logic is needed
to expose the storage location of the RNTuple of the current range, since this
is not available directly from the RPageSource.